save unique queries

+++ app/src/main/java/com/radlance/matule/data/search/ProductSearchRepositoryImpl.kt
@@ -5,6 +5,7 @@ import com.radlance.matule.data.database.local.MatuleDao
 import com.radlance.matule.domain.search.ProductSearchRepository
 import com.radlance.matule.domain.search.SearchHistoryQuery
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
@@ -18,10 +19,15 @@ class ProductSearchRepositoryImpl @Inject constructor(
     }
 
     override suspend fun addQueryToHistory(searchHistoryQuery: SearchHistoryQuery) {
-        dao.insertSearchHistoryQuery(searchHistoryQuery.toSearchHistoryQueryEntity())
-    }
-
-    override suspend fun removeOldestHistoryQuery() {
-        dao.removeOldestHistory()
+        val existingQuery = dao.getSearchHistoryQuery(searchHistoryQuery.query)
+        if (existingQuery != null) {
+            val updatedQuery = existingQuery.copy(queryTime = searchHistoryQuery.queryTime)
+            dao.updateSearchHistoryQuery(updatedQuery)
+        } else {
+            if(dao.getHistory().first().size >= 6) {
+                dao.removeOldestHistory()
+            }
+            dao.insertSearchHistoryQuery(searchHistoryQuery.toSearchHistoryQueryEntity())
+        }
     }
 }
